chore(store): split classes into files and declare store as a module with default export
