+++ drafter/store/__init__.py
@@ -1,0 +1,1 @@
+from drafter.store.main import store
