+++ drafter/store/main.py
@@ -1,0 +1,5 @@
+from drafter.store.Store import Store
+
+# This is the store object that is used to store the state of the application.
+# It is true that it's a global variable, but it stores everything, lowering the complexity.
+store = Store()
